market_predictor/requirements.txt
# Core dependencies
pandas>=1.5.3
numpy>=1.23.5
scikit-learn>=1.2.2
python-dateutil>=2.8.2

# Data handling and storage
pyarrow>=12.0.1  # for parquet support
pyyaml>=6.0.1    # for configuration files

# Market Data
yfinance>=0.2.18  # for market data
fredapi>=0.5.0    # for economic data

# Sentiment Analysis
textblob>=0.15.3
newspaper3k>=0.2.8
beautifulsoup4>=4.9.3
requests>=2.28.2
ratelimit>=2.2.1
websockets>=10.0

# Utilities
python-dotenv>=0.19.2  # for environment variables
logging>=0.5.1.2       # for logging

# Optional but recommended
jupyter>=1.0.0         # for notebooks
ipykernel>=6.21.1      # for notebook kernel